\name{deltaPlotR-package}
\alias{deltaPlotR-package}
\alias{deltaPlotR}
\docType{package}
\title{Identification of dichotomous differential item functioning (DIF) using Angoff's Delta Plot method}


\description{
 The deltaPlotR package proposes an implementation of Angoff's   Delta Plot method to detect dichotomous DIF. Several detection thresholds are included, either from multivariate normality asumption, random simulation, or by prior determination. Item purification is supported.

 }

\details{
 \tabular{ll}{
  Package: \tab deltaPlotR \cr
  Type:    \tab Package \cr
  Version: \tab 1.5 \cr
  Date:    \tab 2014-11-20\cr
  Depends: \tab R (>= 2.10.0), MASS \cr
  License: \tab GPL (>= 2) \cr
  }
 }


\author{
    David Magis \cr
    Post-doc Fellow of the National Funds for Scientific Research (FNRS, Belgium) \cr
    University of Liege \cr
    \email{David.Magis@ulg.ac.be}, \url{http://ppw.kuleuven.be/okp/home/} \cr
    Bruno Facon\cr
    Professor, Department of Psychology \cr
    Universite Lille-Nord de France\cr
    \email{bruno.facon@univ-lille3.fr}, \cr 
    }


\seealso{
 Other useful packages can be found in the \emph{R} \emph{Psychometric} task view.
 See \url{http://cran.stat.sfu.ca/web/views/Psychometrics.html} for further details.
 }


\keyword{package}



